AuthorServiceImpl can throw exceptions now

+++ src/main/java/r4v88/service/AuthorServiceImpl.java
@@ -34,6 +34,8 @@ public class AuthorServiceImpl implements AuthorService {
         for (Map.Entry<Long, Author> authorFromMap : idAuthorsMap.entrySet()) {
             if (authorFromMap.getKey() == id) {
                 author = authorFromMap.getValue();
+            } else {
+                throw new RuntimeException("Author with id: " + id + " does not exist!");
             }
         }
         return author;
@@ -45,6 +47,8 @@ public class AuthorServiceImpl implements AuthorService {
         for (Map.Entry<Long, Author> authorFromMap : idAuthorsMap.entrySet()) {
             if (authorFromMap.getValue().getName().equals(name)) {
                 author = authorFromMap.getValue();
+            } else {
+                throw new RuntimeException("Author with name: " + name + " does not exist!");
             }
         }
         return author;
@@ -56,6 +60,8 @@ public class AuthorServiceImpl implements AuthorService {
         for (Map.Entry<Long, Author> authorFromMap : idAuthorsMap.entrySet()) {
             if (authorFromMap.getValue().getLastname().equals(lastname)) {
                 author = authorFromMap.getValue();
+            } else {
+                throw new RuntimeException("Author with lastname: " + lastname + " does not exist!");
             }
         }
         return author;
@@ -63,11 +69,23 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public void addAuthor(Author author) {
-        authorDao.addAuthor(author);
+        for (Map.Entry<Long, Author> authorFromMap : idAuthorsMap.entrySet()) {
+            if(!authorFromMap.getValue().equals(author)) {
+                authorDao.addAuthor(author);
+            } else {
+                throw new RuntimeException("Author: " + author.getName() + " already exist!");
+            }
+        }
     }
 
     @Override
-    public void removeAuthorById(Long id) {
-        authorDao.removeAuthorById(id);
+    public void removeAuthorById(long id) {
+        for (Map.Entry<Long, Author> authorFromMap : idAuthorsMap.entrySet()) {
+            if (authorFromMap.getKey() == id) {
+                authorDao.removeAuthorById(id);
+            } else {
+                throw new RuntimeException("Author with id: " + id + " does not exist!");
+            }
+        }
     }
 }
